Improve test running display with test count and filenames
Added an header, footer and the number of test that run
Tests are now stripped of the "test_" prefix along with the
filename in which they are located

A new concept has been introduced: the REGISTRY.
This is named after ASM registries and serve as a standard
registry of values for the tests to write in that the test runner
can read and interpret. The first of these value is the test count
that just ran.  Future addition will probably be failures, mock
expectations and such things that should be global to the
test run