Support successive mockings for the same command
By introducing a "workspace" for the mock we can have several
"workspace" related values in files that describe the mock, such
as the successive invocations.

So now the code that is in the fake executable becomes standard
and react to the content of the workspace.

The support for --with-args have been completely removed and
will be reintroduced in an upcoming change with tests

The file sort is to ensure the tests runs the same on any machine,
it would seem that "find" does not guarantee an order