Add cat as a mock option
Currently, mocking executions is limited to echo and exit codes. This
forces test fixtures through test code via cat and echo and sometimes
mangles newlines.

This commit allows the use of cat to point to a file on the filesystem.
This is epecially useful when using $TEST_ROOT_DIR to point to test
fixtures.