Allow running "inplace"
Currently, sources and tests files are copied to the workspace and
executed from there. This works well to prevent accidental files and
directories from being created around the test subject.

However, this means it is impossible to use resources around the test,
without resorting to ${TEST_ROOT_DIR} and, most importantly, tools such
as bashcov are unable to report usage of source files because they have
been run from an external copy.